docker/requirements-ocr.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
python-magic==0.4.27
